move all functors to functors.hpp
