+++ modules/dnn/src/cuda/bias_activation.cu
@@ -5,7 +5,7 @@
 #include <cuda_runtime.h>
 #include <cuda_fp16.h>
 
-#include "activations.hpp"
+#include "functors.hpp"
 #include "types.hpp"
 #include "vector_traits.hpp"
 #include "grid_stride_range.hpp"
